+++ test/unit/fuzz-tools.test.js
@@ -119,6 +119,7 @@ import * as neuroV118 from '../../lib/neuro-v118.js';
 import * as neuroV119 from '../../lib/neuro-v119.js';
 import * as neuroV120 from '../../lib/neuro-v120.js';
 import * as neuroV121 from '../../lib/neuro-v121.js';
+import * as neuroV122 from '../../lib/neuro-v122.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -182,6 +183,7 @@ const MODULES = {
   'neuro-v119.js': neuroV119,
   'neuro-v120.js': neuroV120,
   'neuro-v121.js': neuroV121,
+  'neuro-v122.js': neuroV122,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];